User can now see a sump up of the order

+++ src/main/java/com/rizomm/misys/service/OrderService.java
@@ -18,5 +18,5 @@ public interface OrderService {
 
     RegisteredOrder save();
 
-    Optional<RegisteredOrder> getRegisteredOrderById(long id);
+    Optional<RegisteredOrder> getRegisteredOrderById(Integer id);
 }
